fix(zetasqlite): align BIT_RIGHT_SHIFT with BigQuery's logical right shift (closes #113) (#115)
BigQuery's `>>` is a logical (unsigned) right shift — the upper bits are
zero-filled regardless of sign. Go's int64 `>>` is arithmetic (signed)
and sign-extends, so the emulator diverged from real BigQuery whenever
the base operand was negative.

Verified against the real BigQuery API (project sandbox-248114):
- SELECT -8 >> 2                          → 4611686018427387902 (was -2)
- SELECT CAST(-1 AS INT64) >> 1           → 9223372036854775807 (was -1)
- SELECT -9223372036854775808 >> 1        → 4611686018427387904 (was -4611686018427387904)

BIT_LEFT_SHIFT is unaffected: `<<` produces the same bit pattern in both
signed and unsigned interpretations.

Tests cover both the SQL-operator interpretation layer (query_test.go)
and the Go-function contract layer (function_bit_test.go, new file
named per the function_timestamp_test.go domain-split convention).

+++ internal/zetasqlite/function.go
@@ -111,7 +111,11 @@ func BIT_RIGHT_SHIFT(a, b Value) (Value, error) {
 	if vb < 0 {
 		return nil, fmt.Errorf("Bitwise shift by negative offset.")
 	}
-	return IntValue(va >> vb), nil
+	// BigQuery's `>>` is a *logical* (unsigned) right shift — the upper bits
+	// are zero-filled regardless of sign. Go's int64 `>>` is arithmetic
+	// (signed) and sign-extends, so the negative-base result diverges from
+	// BigQuery. Reinterpret as uint64 to match BigQuery semantics.
+	return IntValue(int64(uint64(va) >> vb)), nil
 }
 
 func BIT_AND(a, b Value) (Value, error) {
